Check if related is authorized in show_relationship
Checked if owner could be seen, but previously skipped over the
related's association ascent because
`Authorization.filter_associatons_can` assumes the related item is
already checked and relate's associations are what is being filtered.

This was probably missed because InterpreterProfile's Address being
private to the owning user is the first show_relationship test case.