Standardize dasherizing relationship names
Only one view was dasherizing its relationship, this lead to the weird
whack-a-mole with the formatting of the include keys in
`Calcinator.JaSerializer.PhoenixView`.  Naming the relationships dasherized as
the benefit that the params don't need to be reformatted just for views and
they can be passed to other layers that expect the JSONAPI format.