Wrap Controller.* in Code.ensure_loaded?
`Code.ensure_loaded?(Phoenix.Controller)` can be used to protect
`Calcinator.Controller.Error` and `Calcinator.Controller`, so that it is not
defined when its dependency, `Phoenix.Controller` is not available.  Without
this change, `(CompileError) lib/calcinator/controller/error.ex:12: module
Phoenix.Controller is not loaded and could not be found` is raised in
retort.