Configurable force pagination
Instead of having :default_page_size, there is now :page_size, which is
a Keyword.t in `config` or `Application.get_env` and and map in the
code.  It has keys `:default` to replace `:default_page_size`;
`:minimum` for the minimum allowed page size; and `:maximum` for the
maximum allowed page size.  If either `:minimum` or `:maximum` is
configured, then pagination is forced and an error will be returned if
`%{page: nil}` is passed in the `query_options` to try to disable
default pagination.  The `:default` does not need to be specified if
either `:minimum` or `:maximum` is specified: when `:maximum` is
present, it will be used as the default; when only `:minimum` is
present, it will be the default; if no key is present, then there is no
default pagination, but it is still allowed through `query_options`.