allow_sandbox_access before getting changeset
`Calcinator.Resources.changeset/1,2` is allowed to get `many_to_many`
associations from the backing store, so during testing, this means that
the sandbox access must be allowed prior to calling `changeset`.  This
was already the case for other actions, but for `create`,
`allow_sandbox_access` was not called until the changeset was authorized
and about to be created.